Refactor gmute and score module
